add new proposer framework, add mixture proposer
