+++ LOTlib/Hypotheses/Proposers/__init__.py
@@ -1,10 +1,7 @@
-class ProposalFailedException(Exception):
-    """
-    This gets raised when we have a proposal that can't succeed
-    """
-    pass
-
-from CopyRegenProposal import copy_regen_proposal
-from InsertDeleteProposal import insert_delete_proposal
-from RegenerationProposal import regeneration_proposal
-from CopyRegenProposal import copy_proposal
+#from CopyRegenProposal import copy_regen_proposal
+#from InsertDeleteProposal import insert_delete_proposal
+from RegenerationProposer import RegenerationProposer
+regeneration_proposal = RegenerationProposer().proposal_content
+from CopyProposer import CopyProposer
+copy_proposal = CopyProposer().proposal_content
+from MixtureProposer import MixtureProposer
